correções conforme pr
