+++ features/favorite/src/androidTest/java/com/example/favorite/di/TestModule.kt
@@ -6,9 +6,7 @@ import org.koin.core.context.loadKoinModules
 import org.koin.dsl.module
 
 internal class TestModule {
-
     private val modules = module {
-
         factory { FavoriteFragmentRobot() }
         factory { FileReader() }
     }
